Added variant visit library
This is based on experiments compiler explorer, after discovering that std::visit did not produce optimal code on most compilers, when visiting std::variant. When traversing an array, there were many more jumps and calls than expected. Clang did a little better, but as soon as the array could be changed at runtime, again many calls were generated, along with vtables which might relate to exceptions. This is a first draft which accepts a visitor with specialised call operators() that take no arguments.